// src/app/shared/api/search/models/model.ts
export interface Model {
  id: number;
  projectId?: number;
  area: number;
  deliveryDate: string;
  projectName?: string;
  building: string;
  numberOfBedrooms?: string;
  numberOfBathrooms?: string;
  usageTypeName: string;
  usageTypeId?: string;
  modelCode?: string;
  floor?: string;
  modelImageUrl?: string;
  modelId: string;
  projectCode?: string;
}